+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 are in a single HTTP session.
 - `conformance.CONFORMANCE_CLASSES` dictionary added containing all STAC API Capabilities from stac-api-spec
 - `collections` subcommand to CLI, for saving all Collections in catalog as JSON
+- `Client.get_collections` overrides Catalog to use /collections endpoint if API conforms
 
 ### Changed
 
@@ -42,6 +43,7 @@ are in a single HTTP session.
 - `conformance.ConformanceClass`, replaced with `conformance.ConformanceMixin`
 - `conformance.ConformanceClasses`, replaced with `conformance.ConformanceMixin`
 - PySTAC Collection objects can no longer be passed in as `collections` arguments to the `ItemSearch` class (just pass ids)
+- Catalog.get_collection_list (was alias to get_child_links) because made assumption about this being an API only. Also redundant with `Catalog.get_collections`
 
 ### Deprecated
 - `Search.item_collections()`